backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
pydantic
fastmcp[cli]
langchain-mcp 